Refactors demo code so it can be used for multiple use cases
