+++ demos/basic/index.html
@@ -2,14 +2,17 @@
 <html lang="en">
 <head>
 	<meta charset="UTF-8">
-	<base href="/demos/basic/">
+	<base href="/demos/">
 	<title>Demo: Metal Router</title>
 
-	<link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.css">
-	<link rel="stylesheet" href="../../node_modules/senna/build/senna.css">
+	<link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.css">
+	<link rel="stylesheet" href="../node_modules/senna/build/senna.css">
 </head>
 <body>
 	<div class="senna-loading-bar"></div>
 	<script src="build/metal.js"></script>
+	<script>
+		metal.Basic.run();
+	</script>
 </body>
 </html>
